added authentication to the apis

+++ traversy media MERN course/api/controllers/goalController.js
@@ -1,8 +1,9 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
   res.status(200).json(goals);
 });
 
@@ -13,6 +14,7 @@ const postGoals = asyncHandler(async (req, res) => {
   }
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user.id,
   });
   res.status(201).json(goal);
 });
@@ -21,8 +23,22 @@ const updateGoals = asyncHandler(async (req, res) => {
   const goal = await Goal.findById(req.params.id);
   if (!goal) {
     res.status(400);
-    throw new Error("Goal not found");
+    throw new Error("User not found");
   }
+
+  const user = await User.findById(req.user.id);
+
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  // checking logged in user with goal's user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorizes");
+  }
+
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
@@ -30,6 +46,18 @@ const updateGoals = asyncHandler(async (req, res) => {
 });
 
 const deleteGoals = asyncHandler(async (req, res) => {
+  const user = await User.findById(req.user.id);
+
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorizes");
+  }
+
   const goal = await Goal.findByIdAndRemove(req.params.id);
   if (!goal) {
     res.status(400);
